refactore(theme) creando el archivo theme ,  moviendo colores y el main theme aqui close #36
