+++ packages/task_repository/lib/src/models/models.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'package:equatable/equatable.dart';
 import 'package:formz/formz.dart';
 import 'package:oshin_list/helper/color_helper.dart';
+import 'package:oshin_list/utils/theme.dart';
 
 part 'task.dart';
 part 'type.dart';
